fix: room.ts로 코드 이동 및 방장이 방 삭제하는 로직 추가
