+++ src/api/supabaseCSR/supabase.ts
@@ -212,19 +212,3 @@ export const deleteSelectedDate = async ({ roomId, userId, date }: { roomId: str
 
   if (error) throw error;
 };
-
-export const deleteExitUserData = async (payload: { roomId: string; userId: string }) => {
-  const { error } = await supabase.from('userdata_room').delete().eq('user_id', payload.userId);
-
-  if (error) throw error;
-};
-
-export const deleteExitUserSchedule = async (payload: { roomId: string; userId: string }) => {
-  const { error } = await supabase
-    .from('room_schedule')
-    .delete()
-    .eq('room_id', payload.roomId)
-    .eq('created_by', payload.userId);
-
-  if (error) throw error;
-};
